% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_letters}
\alias{sample_letters}
\title{Sample letters suited for textured package}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
sample_letters
}
\description{
A dataset containing two sample letters, in a format best suited for
the textured functions to operate on.
}
\keyword{datasets}
