% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_create_element.R
\name{txx_create_element}
\alias{txx_create_element}
\title{Create an XML element from a string}
\usage{
txx_create_element(string, tag)
}
\arguments{
\item{string}{}

\item{tag}{}
}
\value{

}
\description{
Create an XML element from a string
}
\examples{

}
