% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_finish_sections.R
\name{txx_finish_sections}
\alias{txx_finish_sections}
\title{Adds end tags to match start tags.}
\usage{
txx_finish_sections(string, split = "¬")
}
\arguments{
\item{string}{The character string containing a single entry. Should have
already been through the xml-transforming process so should have start
sections identified and wrapped into a start tag.}

\item{split}{The character that is used to split the string}
}
\value{
A character string; a modified of \code{string}
}
\description{
Splits a string into the (incomplete) tagged sections according to the split 
character, then identifies the start tag and adds a corresponding end tag,
and then recombines it into a single string.
}
\examples{



}
