% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_identify_sections.R
\name{txx_identify_sections}
\alias{txx_identify_sections}
\title{Identify sections within text, and wrap in tags}
\usage{
txx_identify_sections(string, tags, remove = NA_character_, ignore_case = TRUE)
}
\arguments{
\item{string}{A character string of the document to be searched}

\item{tags}{A vector of character strings - strings used to identify a 
section of text}

\item{ignore_case}{Logic value of whether you want to ignore the case when
identifying the sections}
}
\value{
A character string with sections identified and wrapped in an XML tag.
}
\description{
This function identifies the tags within a character string, and then wraps
the identified text in XML tag syntax. Generally not used outside of the
\code{txx_text_to_xml} function, but has its uses.
}
\details{
Note that tag order is important - the first tags are identified first and
modify the string. If there is another tag that would have identified a
section within that string before the first tag, it would not identify it
after.

Tags can be regular expressions - use with caution to avoid mis-identifying
sections.
}
\examples{


}
