% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_sanitize_tag_name.R
\name{txx_sanitize_tag_name}
\alias{txx_sanitize_tag_name}
\title{Sanitize text for use as an XML tag}
\usage{
txx_sanitize_tag_name(tag_name)
}
\arguments{
\item{tag_name}{A character string}
}
\value{
The character string, modified to be suitable for use as an XML tag
}
\description{
This function takes a character string, and returns a 'sanitized' version
of it for use as an XML tag (following XML specifications - more details 
below). This function is called internally, however is useful in searching
for a specific text tag(s).
}
\details{
Changes made to the text (in this order): 
\enumerate{
   \item Trims any white space around the tag, 
   \item Changes text to title case, 
   \item Replaces all ' ' with '_', 
   \item Removes any non-alphanumeric characters (aside from '_'), 
   \item Prefixes a '_' if the text starts with a number.
}
}
\examples{

}
