% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_sanitize_tag_sections.R
\name{txx_sanitize_tag_sections}
\alias{txx_sanitize_tag_sections}
\title{Sanitize a tag section within an XML document}
\usage{
txx_sanitize_tag_sections(string)
}
\arguments{
\item{string}{String consisting of \code{<tag>section}}
}
\value{
String with \code{section} sanitized for use in an XML document
}
\description{
Identifies the section within a tag (actually after a tag once split), and
sanitizes the text of the section (see \code{txx_sanitize_text()})
}
\examples{


}
