% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_sanitize_text.R
\name{txx_sanitize_text}
\alias{txx_sanitize_text}
\title{Sanitize text to be stored as data in an XML document}
\usage{
txx_sanitize_text(text)
}
\arguments{
\item{text}{A character string}
}
\value{
A modified character string
}
\description{
This function takes a character string and returns a 'sanitized' version,
suitable for storage as text within an XML document.
}
\details{
This function replaces any special characters within XML, and replaces them
with the relevant XML escaped characters. As they are read in as XML, they
should be transformed back into the proper version.
}
\examples{
txx_sanitize_text("Alice & Bob have been in \"Charlie's\" house for >5 minutes.")

}
