% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_wrap_tag.R
\name{txx_wrap_tag}
\alias{txx_wrap_tag}
\title{Turn text into an XML tag}
\usage{
txx_wrap_tag(tag, split_character = "¬", start = "<", end = ">")
}
\arguments{
\item{tag}{The text to turn into a tag}

\item{split_character}{For use in the structuring process. Should be a 
character that is not seen in the document.}

\item{start}{The first character to wrap the text in}

\item{end}{The last character to wrap the text in}
}
\value{
A character string of the text as an XML tag
}
\description{
Turns some text into an XML tag sanitizing the text to be suitable for
use as an XML tag, then wrapping it in \code{<} and \code{>}.
}
