% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_create_element.R
\name{txx_create_element}
\alias{txx_create_element}
\title{Create an XML element from a string}
\usage{
txx_create_element(string, tag)
}
\arguments{
\item{string}{A character string}

\item{tag}{The character string used in the start and end tags.}
}
\value{
A character string
}
\description{
Converts a string into an XML element by adding start and end tags. The
string can also be an XML element, allowing for easy nesting of XML
elements.
}
\examples{

}
