% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_sanitize_sections.R
\name{txx_sanitize_sections}
\alias{txx_sanitize_sections}
\title{Sanitizes the text as part of the txx_text_to_xml process}
\usage{
txx_sanitize_sections(string, split = "¬")
}
\description{
Sanitizes the text as part of the txx_text_to_xml process
}
