% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_sanitize_sections.R
\name{txx_sanitize_sections}
\alias{txx_sanitize_sections}
\title{Title}
\usage{
txx_sanitize_sections(string, split = "¬")
}
\arguments{
\item{string}{}

\item{split}{}
}
\value{

}
\description{
Title
}
