% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txx_text_to_xml.R
\name{txx_text_to_xml}
\alias{txx_text_to_xml}
\title{Test test test}
\usage{
txx_text_to_xml(strings, tags, remove = ":")
}
\arguments{
\item{strings}{}

\item{tags}{}

\item{remove}{}
}
\value{

}
\description{
test test test
}
\examples{


}
